Breaking down new account into 2 separate scripts
