+++ .gitignore
@@ -139,3 +139,6 @@ cython_debug/
 
 # IDEs
 .idea/
+
+# ivy-account-tools specific
+setup_sso/*.xml
