+++ setup_sso/README.md
@@ -1,0 +1,24 @@
+# Sets up SAML on an AWS Account
+
+## How to use it
+
+```
+$ cd path/to/this/repo
+$ AWS_PROFILE=profile-to-use pipenv run python setup_sso/setup_sso.py -h
+usage: setup_sso.py [-h] -a SUB_ACCOUNT_NAME [-f SAML_FILE] [-s SAML_PROVIDER] [-t IVY_TAG] [-l {CRITICAL,ERROR,WARNING,INFO,DEBUG}]
+
+Sets up an AWS account's alias, SAML provider, Admin role and Read-Only role
+
+optional arguments:
+  -h, --help            show this help message and exit
+  -a SUB_ACCOUNT_NAME, --sub-account-name SUB_ACCOUNT_NAME
+                        AWS Sub Account Name
+  -f SAML_FILE, --saml-metadata-document-file SAML_FILE
+                        Path to An XML document generated by an identity provider (IdP) that supports SAML 2.0
+  -s SAML_PROVIDER, --saml-provider-name SAML_PROVIDER
+                        Name of the saml provider. Examples: gsuite, msft
+  -t IVY_TAG, --ivy-tag IVY_TAG
+                        Ivy tag also known as namespace
+  -l {CRITICAL,ERROR,WARNING,INFO,DEBUG}, --log-level {CRITICAL,ERROR,WARNING,INFO,DEBUG}
+                        Set the logging output level
+```
